+++ lib/hanami/helpers/form_helper/form_builder.rb
@@ -83,11 +83,11 @@ module Hanami
 
         # Instantiate a form builder
         #
-        # @overload initialize(form, attributes, params, &blk)
+        # @overload initialize(form, attributes, context, &blk)
         #   Top level form
         #   @param form [Hanami::Helpers:FormHelper::Form] the form
         #   @param attributes [::Hash] a set of HTML attributes
-        #   @param params [Hanami::Action::Params] request params
+        #   @param context [Hanami::Helpers::FormHelper]
         #   @param blk [Proc] a block that describes the contents of the form
         #
         # @overload initialize(form, attributes, params, &blk)
@@ -150,6 +150,9 @@ module Hanami
         # @param name [Symbol] the nested name, it's used to generate input
         #   names, ids, and to lookup params to fill values.
         #
+        # @yield [index]
+        # @yieldparam [Integer] index iterative index (it starts from zero)
+        #
         # @since 0.2.0
         #
         # @example Basic usage
@@ -202,10 +205,10 @@ module Hanami
         #
         #     <button type="submit">Create</button>
         #   </form>
-        def fields_for(name)
+        def fields_for(name, value = nil)
           current_name = @name
           @name        = _input_name(name)
-          yield(name)
+          yield(name, value)
         ensure
           @name = current_name
         end
@@ -217,6 +220,10 @@ module Hanami
         #
         # @param name [Symbol] the nested name, it's used to generate input
         #   names, ids, and to lookup params to fill values.
+        #
+        # @yield [index, value]
+        # @yieldparam [Integer] index iterative index (it starts from zero)
+        # @yieldparam [Object] value an item of the collection
         #
         # @example Basic usage
         #   <%=
@@ -240,13 +247,60 @@ module Hanami
         #
         #     <button type="submit">Create</button>
         #   </form>
+        #
+        # @example Yield index and value
+        #   <%=
+        #     form_for(:bill, routes.bill_path(id: bill.id), { bill: bill }, method: :patch, class: 'form-horizontal') do
+        #       fieldset do
+        #         legend "Addresses"
+        #
+        #         fields_for_collection :addresses do |i, address|
+        #           div class: "form-group" do
+        #             text "Address id: #{address.id}"
+        #
+        #             label :street
+        #             input_text :street, class: "form-control", placeholder: "Street", "data-funky": "index-#{i}"
+        #           end
+        #         end
+        #
+        #         label :ensure_names
+        #       end
+        #
+        #       submit submit_label, class: "btn btn-default"
+        #     end
+        #   %>
+        #
+        #   <!-- output -->
+        #   <form action="/bills/1" method="POST" accept-charset="utf-8" id="bill-form" class="form-horizontal">
+        #     <input type="hidden" name="_method" value="PATCH">
+        #     <input type="hidden" name="_csrf_token" value="920cd5bfaecc6e58368950e790f2f7b4e5561eeeab230aa1b7de1b1f40ea7d5d">
+        #     <fieldset>
+        #       <legend>Addresses</legend>
+        #
+        #       <div class="form-group">
+        #         Address id: 23
+        #         <label for="bill-addresses-0-street">Street</label>
+        #         <input type="text" name="bill[addresses][][street]" id="bill-addresses-0-street" value="5th Ave" class="form-control" placeholder="Street" data-funky="index-0">
+        #       </div>
+        #
+        #       <div class="form-group">
+        #         Address id: 42
+        #         <label for="bill-addresses-1-street">Street</label>
+        #         <input type="text" name="bill[addresses][][street]" id="bill-addresses-1-street" value="4th Ave" class="form-control" placeholder="Street" data-funky="index-1">
+        #       </div>
+        #
+        #       <label for="bill-ensure-names">Ensure names</label>
+        #     </fieldset>
+        #
+        #     <button type="submit" class="btn btn-default">Update</button>
+        #   </form>
         def fields_for_collection(name, &block)
           current_name = @name
           base_value = _value(name)
           @name = _input_name(name)
 
-          base_value.count.times do |index|
-            fields_for(index, &block)
+          base_value.each_with_index do |value, index|
+            fields_for(index, value, &block)
           end
         ensure
           @name = current_name
@@ -1227,17 +1281,19 @@ module Hanami
         #     <option value="zw">Zimbabwe</option>
         #   </select>
         def select(name, values, attributes = {}) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
-          options    = attributes.delete(:options) { {} }
-          attributes = { name: _select_input_name(name, attributes[:multiple]), id: _input_id(name) }.merge(attributes)
-          prompt     = options.delete(:prompt)
-          selected   = options.delete(:selected)
+          options     = attributes.delete(:options) { {} }
+          multiple    = attributes[:multiple]
+          attributes  = { name: _select_input_name(name, multiple), id: _input_id(name) }.merge(attributes)
+          prompt      = options.delete(:prompt)
+          selected    = options.delete(:selected)
+          input_value = _value(name)
 
           super(attributes) do
             option(prompt) unless prompt.nil?
 
             already_selected = nil
             values.each do |content, value|
-              if (attributes[:multiple] || !already_selected) && (already_selected = _select_option_selected?(value, selected, _value(name), attributes[:multiple]))
+              if (multiple || !already_selected) && (already_selected = _select_option_selected?(value, selected, input_value, multiple))
                 option(content, { value: value, selected: SELECTED }.merge(options))
               else
                 option(content, { value: value }.merge(options))
@@ -1328,8 +1384,15 @@ module Hanami
 
         # Button
         #
-        # @param content [String] The content
-        # @param attributes [Hash] HTML attributes to pass to the button tag
+        # @overload button(content, attributes = {})
+        #   Use string content
+        #   @param content [String] The content
+        #   @param attributes [Hash] HTML attributes to pass to the button tag
+        #
+        # @overload button(attributes = {}, &blk)
+        #   Use block content
+        #   @param attributes [Hash] HTML attributes to pass to the button tag
+        #   @param blk [Proc] the block content
         #
         # @since 1.0.0
         #
@@ -1350,8 +1413,25 @@ module Hanami
         #
         #   <!-- output -->
         #   <button class="btn btn-secondary">Click me</button>
-        def button(content, attributes = {})
-          # This is here only for documentation purposes
+        #
+        # @example Block
+        #   <%=
+        #     # ...
+        #     button class: "btn btn-secondary" do
+        #       span class: 'oi oi-check'
+        #     end
+        #   %>
+        #
+        #   <!-- output -->
+        #   <button class="btn btn-secondary">
+        #     <span class="oi oi-check"></span>
+        #   </button>
+        def button(content, attributes = {}, &blk)
+          if content.is_a?(::Hash)
+            attributes = content
+            content = nil
+          end
+
           super
         end
 
@@ -1393,8 +1473,15 @@ module Hanami
 
         # Submit button
         #
-        # @param content [String] The content
-        # @param attributes [Hash] HTML attributes to pass to the button tag
+        # @overload submit(content, attributes = {})
+        #   Use string content
+        #   @param content [String] The content
+        #   @param attributes [Hash] HTML attributes to pass to the button tag
+        #
+        # @overload submit(attributes = {}, &blk)
+        #   Use block content
+        #   @param attributes [Hash] HTML attributes to pass to the button tag
+        #   @param blk [Proc] the block content
         #
         # @since 0.2.0
         #
@@ -1415,9 +1502,27 @@ module Hanami
         #
         #   <!-- output -->
         #   <button type="submit" class="btn btn-primary">Create</button>
-        def submit(content, attributes = {})
+        #
+        # @example Block
+        #   <%=
+        #     # ...
+        #     button class: "btn btn-primary" do
+        #       span class: 'oi oi-check'
+        #     end
+        #   %>
+        #
+        #   <!-- output -->
+        #   <button type="submit" class="btn btn-primary">
+        #     <span class="oi oi-check"></span>
+        #   </button>
+        def submit(content, attributes = {}, &blk)
+          if content.is_a?(::Hash)
+            attributes = content
+            content = nil
+          end
+
           attributes = { type: :submit }.merge(attributes)
-          button(content, attributes)
+          button(content, attributes, &blk)
         end
 
         protected
@@ -1570,20 +1675,38 @@ module Hanami
           select_name
         end
 
-        # TODO: this has to be refactored
-        #
         # @api private
-        #
-        # rubocop:disable Metrics/CyclomaticComplexity
-        # rubocop:disable Metrics/PerceivedComplexity
         def _select_option_selected?(value, selected, input_value, multiple)
-          value == selected ||
-            (multiple && (selected.is_a?(Array) && selected.include?(value))) ||
-            (!input_value.nil? && (value.to_s == input_value.to_s)) ||
-            (multiple && (input_value.is_a?(Array) && input_value.include?(value)))
+          if input_value && selected.nil?
+            value.to_s == input_value
+          else
+            (value == selected) ||
+              _is_in_selected_values?(multiple, selected, value) ||
+              _is_current_value?(input_value, value) ||
+              _is_in_input_values?(multiple, input_value, value)
+          end
         end
-        # rubocop:enable Metrics/PerceivedComplexity
-        # rubocop:enable Metrics/CyclomaticComplexity
+
+        # @api private
+        # @since 1.2.0
+        def _is_current_value?(input_value, value)
+          return unless input_value
+          value.to_s == input_value.to_s
+        end
+
+        # @api private
+        # @since 1.2.0
+        def _is_in_selected_values?(multiple, selected, value)
+          return unless multiple && selected.is_a?(Array)
+          selected.include?(value)
+        end
+
+        # @api private
+        # @since 1.2.0
+        def _is_in_input_values?(multiple, input_value, value)
+          return unless multiple && input_value.is_a?(Array)
+          input_value.include?(value)
+        end
 
         # @api private
         def _check_box_checked?(value, input_value)