+++ spec/support/fixtures.rb
@@ -193,7 +193,7 @@ class NumbersView
   end
 
   def big_decimal
-    format_number BigDecimal.new("0.0001"), precision: 4
+    format_number BigDecimal("0.0001"), precision: 4
   end
 
   def complex
@@ -379,9 +379,10 @@ class SessionFormHelperView < FormHelperView
 end
 
 class Address
-  attr_reader :street
+  attr_reader :id, :street
 
   def initialize(attributes = {})
+    @id     = attributes[:id]
     @street = attributes[:street]
   end
 end
@@ -448,6 +449,10 @@ module FullStack
       _escape "/sessions"
     end
 
+    def settings_path
+      _escape "/settings"
+    end
+
     def deliveries_path
       _escape "/deliveries"
     end
@@ -488,6 +493,17 @@ module FullStack
       class New
         include TestView
         template "sessions/new"
+      end
+    end
+
+    module Settings
+      class Edit
+        include TestView
+        template "settings/edit"
+
+        def form
+          Form.new(:settings, routes.settings_path)
+        end
       end
     end
 
@@ -536,6 +552,19 @@ module FullStack
       class Edit
         include TestView
         template "bills/edit"
+
+        def form
+          Form.new(:bill, routes.bill_path(id: bill.id), { bill: bill }, method: :patch)
+        end
+
+        def submit_label
+          "Update"
+        end
+      end
+
+      class Edit2
+        include TestView
+        template "bills/edit2"
 
         def form
           Form.new(:bill, routes.bill_path(id: bill.id), { bill: bill }, method: :patch)