+++ spec/unit/hanami/helpers/html_helper/html_builder_spec.rb
@@ -247,6 +247,18 @@ CONTENT
       result = @builder.input("required" => true, "value" => 'Title "book"', "something" => "bar").to_s
       expect(result).to eq('<input required="required" value="Title "book"" something="bar">')
     end
+
+    # rubocop:disable Style/SymbolArray
+    it "renders empty node with array value as joined with space" do
+      result = @builder.input("class" => [:ui, :form]).to_s
+      expect(result).to eq('<input class="ui form">')
+    end
+
+    it "renders non-empty node with array value as joined with space" do
+      result = @builder.span("foo", "class" => [:ui, :form]).to_s
+      expect(result).to eq('<span class="ui form">foo</span>')
+    end
+    # rubocop:enable Style/SymbolArray
   end
 
   ##############################################################################